// services/equipment-service/src/domain/constants/enums.ts
export enum EEquipmentTypes {
  Amulet = "Amulet",
  Belt = "Belt",
  Body = "Body",
  Boots = "Boots",
  Gloves = "Gloves",
  Helmet = "Helmet",
  Legs = "Legs",
  Relic = "Relic",
  Ring = "Ring",
  Shield = "Shield",
  Sword = "Sword",
  Cloak = "Cloak",
  Potion = "Potion",
  Book = "Book",
  Toolkit = "Toolkit",
}

export enum EArmourTypes {
  Cloth = "Cloth",
  Light = "Light",
  Medium = "Medium",
  Heavy = "Heavy",
}

export enum EDamageTypes {
  Piercing = "Piercing",
  Slashing = "Slashing",
  Bludgeoning = "Bludgeoning",
  Acid = "Acid",
  Cold = "Cold",
  Fire = "Fire",
  Force = "Force",
  Lightning = "Lightning",
  Necrotic = "Necrotic",
  Poison = "Poison",
  Psychic = "Psychic",
  Radiant = "Radiant",
  Thunder = "Thunder",
}

export enum ECharacteristics {
  STR = "STR",
  DEX = "DEX",
  CON = "CON",
  INT = "INT",
  WIS = "WIS",
  CHA = "CHA",
}

export enum EDiceTypes {
  D2 = "d2",
  D4 = "d4",
  D6 = "d6",
  D8 = "d8",
  D10 = "d10",
  D12 = "d12",
  D20 = "d20",
  D100 = "d100",
}